# submissions/track1/utxo-indexer-helios/docs/swagger.yaml
basePath: /
definitions:
  models.MultiSelectRequest:
    properties:
      addresses:
        items:
          type: string
        minItems: 1
        type: array
      amount:
        minimum: 1
        type: integer
      fee_rate:
        description: satoshis per byte
        type: integer
      strategy:
        $ref: '#/definitions/models.SelectStrategy'
    required:
    - addresses
    - amount
    type: object
  models.SelectStrategy:
    enum:
    - optimize_fee
    - optimize_fee_1
    - optimize_fee_2
    - optimize_fee_3
    - optimize_fee_4
    - optimize_fee_5
    - optimize_fee_6
    - privacy
    - consolidation
    - default
    type: string
    x-enum-comments:
      StrategyOptimizeFee1: Effective Value
      StrategyOptimizeFee2: Branch and Bound
      StrategyOptimizeFee3: Knapsack DP
      StrategyOptimizeFee4: Single Random Draw
      StrategyOptimizeFee5: Accumulative
      StrategyOptimizeFee6: Genetic Algorithm
    x-enum-descriptions:
    - ""
    - Effective Value
    - Branch and Bound
    - Knapsack DP
    - Single Random Draw
    - Accumulative
    - Genetic Algorithm
    - ""
    - ""
    - ""
    x-enum-varnames:
    - StrategyOptimizeFee
    - StrategyOptimizeFee1
    - StrategyOptimizeFee2
    - StrategyOptimizeFee3
    - StrategyOptimizeFee4
    - StrategyOptimizeFee5
    - StrategyOptimizeFee6
    - StrategyPrivacy
    - StrategyConsolidation
    - StrategyDefault
  models.SelectUTXORequest:
    properties:
      address:
        type: string
      amount:
        minimum: 1
        type: integer
      fee_rate:
        description: satoshis per byte
        type: integer
      strategy:
        $ref: '#/definitions/models.SelectStrategy'
    required:
    - address
    - amount
    type: object
  models.SelectUTXOResponse:
    properties:
      change:
        type: integer
      estimated_fee:
        type: integer
      num_inputs:
        type: integer
      selected_utxos:
        items:
          $ref: '#/definitions/models.UTXO'
        type: array
      total_value:
        type: integer
    type: object
  models.UTXO:
    properties:
      address:
        type: string
      block_height:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      script_pubkey:
        type: string
      spent:
        type: boolean
      spent_at:
        type: string
      spent_txid:
        type: string
      txid:
        type: string
      value:
        type: integer
      vout:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/helios/utxo-indexer
  description: |-
    Bitcoin UTXO management service with 6 advanced selection algorithms and Multi-Address support

    This API provides efficient UTXO indexing and selection for Bitcoin wallets,
    supporting both single and multiple address management with various selection strategies.

    Features:
    - 6 Advanced UTXO selection algorithms (Knapsack DP, Branch & Bound, GA, etc.)
    - Multi-Address support for HD wallets
    - Real-time UTXO management
    - Fee optimization strategies
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: UTXO Indexer-Selector API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Get service information and available endpoints
      produces:
      - application/json
      responses:
        "200":
          description: Service info with endpoints and usage examples
          schema:
            additionalProperties: true
            type: object
      summary: Service information
      tags:
      - Info
  /balance/{address}:
    get:
      consumes:
      - application/json
      description: Get the total balance for a specific Bitcoin address
      parameters:
      - description: Bitcoin address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balance information
          schema:
            additionalProperties: true
            type: object
      summary: Get balance for address
      tags:
      - Balance
  /multi-select:
    post:
      consumes:
      - application/json
      description: Select optimal UTXOs from multiple Bitcoin addresses for a transaction
      parameters:
      - description: Multi-address selection parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MultiSelectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Selected UTXOs from multiple addresses
          schema:
            $ref: '#/definitions/models.SelectUTXOResponse'
        "400":
          description: Bad request or insufficient funds
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Multi-Address UTXO Selection
      tags:
      - Multi-Address
  /multi-summary:
    post:
      consumes:
      - application/json
      description: Get balance and UTXO summary for multiple Bitcoin addresses
      parameters:
      - description: List of addresses
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Balance summary for all addresses
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Multi-Address Balance Summary
      tags:
      - Multi-Address
  /refresh:
    post:
      consumes:
      - application/json
      description: Refresh UTXO data for a specific address from live API
      parameters:
      - description: Address to refresh
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Refresh successful
          schema:
            additionalProperties: true
            type: object
      summary: Refresh UTXOs
      tags:
      - UTXOs
  /select:
    post:
      consumes:
      - application/json
      description: Select optimal UTXOs for a transaction based on amount and strategy
      parameters:
      - description: Selection parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SelectUTXORequest'
      produces:
      - application/json
      responses:
        "200":
          description: Selected UTXOs
          schema:
            $ref: '#/definitions/models.SelectUTXOResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Select UTXOs for transaction
      tags:
      - Selection
  /spend:
    post:
      consumes:
      - application/json
      description: Mark UTXOs as spent (demo only)
      parameters:
      - description: Spend request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Simulate spending
      tags:
      - Demo
  /stats/{address}:
    get:
      consumes:
      - application/json
      description: Get UTXO statistics for a specific Bitcoin address
      parameters:
      - description: Bitcoin address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Statistics
          schema:
            additionalProperties: true
            type: object
      summary: Get statistics for address
      tags:
      - Statistics
  /utxos/{address}:
    get:
      consumes:
      - application/json
      description: Get all unspent transaction outputs for a specific Bitcoin address
      parameters:
      - description: Bitcoin address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of UTXOs
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get UTXOs for address
      tags:
      - UTXOs
schemes:
- http
- https
swagger: "2.0"